requirements.txt
# Core Resume Parsing
azure-ai-formrecognizer
python-dotenv
regex
spacy

# Language Model (OpenAI)
openai

# Web Interface
streamlit

# PDF & DOCX Support
PyPDF2
fpdf
python-docx

en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0.tar.gz